script for remounting storage with exec privileges
When I close my Chromebook, the SD card unmounts. When it auto mounts
again, it doesn't have execution privileges. This is a problem, since I
like to keep bulky software on it, and the command for remounting is a
little long. So, this is basically a fancier alias for that.

It checks if it's already mounted before remounting, so I don't
accidentally interrupt access in case something is using it.

For referece, the command is
sudo mount -o remount,exec /media/removable/evo
which is a fair bit longer than
sudo mkexec ~/SD
.

I also set the backronym script to ignore errors when joining the words,
since I don't care enough to fix it every time the dictionary adds new
wierd characters. (It might be worth while at some point to figure out
how to handle unicode cleanly in Python.)