Made mkexec more user friendly
Because root does not use my own .bashrc, it does not have the same
PATH. So, when putting the sudo outside of mkexec, root can't find
mkexec. So, I put the sudo in with the mount command. This has the added
bonus of only asking me for my password if I actually need it.