various desktop launchers
The Jupyter launcher is a fairly complicated one-liner that checks for
running jupyter servers, attaches to tmux, and opens a browser tab.
If there is no tmux session, it fails. At some point I would like to fix
this. I'll probably need to make a helper script, since it's hard to
find bugs in such a long one-liner.