more physically meaningful prediction method
Previously, the only prediction I displayed was a quadratic fit of the
active cases, extrapolated out one week.

Of course, the "source" of the active cases is the growth rate,
representing how spread-able the virus is at a given time.
So now, I take a quadratic fit of the growth rate and use that to
extrapolate today's active cases forwards and backwards in time, using
Euler's method for numerical integration.

This is more physically meaningful, since the *exponential* growth of
the active cases is driven by the growth rate. (For example, a constant
growth rate would result in an *exponential* growth in the active cases,
not quadratic growth.) Since the growth rate is driven by wierd
socio-political processes, it makes more sense to do the quatratic fit
there.